+++ types/collection.hh
@@ -26,6 +26,7 @@
 #include <vector>
 
 #include "types.hh"
+#include "collection_mutation.hh"
 #include "utils/chunked_vector.hh"
 #include "schema_fwd.hh"
